// server/vance-addon-brain-scribble/client/src/generated/scribble/ScribblebookView.ts
/*
 Auto-generated by generate-java-to-ts-maven-plugin. Do not edit.
 Source: de.mhus.vance.addon.brain.scribble.ScribblebookView
*/

import { ScribblebookPageView } from './ScribblebookPageView';

export interface ScribblebookView {
  folder: string;
  title?: string;
  description?: string;
  landingPagePath?: string;
  landingPageId?: string;
  pages: ScribblebookPageView[];
}