// server/vance-addon-brain-scribble/client/src/generated/scribble/ScribbleSheetView.ts
/*
 Auto-generated by generate-java-to-ts-maven-plugin. Do not edit.
 Source: de.mhus.vance.addon.brain.scribble.ScribbleSheetView
*/

import { ScribbleSheetDto } from './ScribbleSheetDto';

export interface ScribbleSheetView {
  id: string;
  path: string;
  title?: string;
  sheet: ScribbleSheetDto;
}